src/edgemodules/modules/PredictModule/requirements.txt
azure-iot-device~=2.7.0
pillow==8.3.1
tensorflow==2.6